Moved pkg/proto and pkg/sched/ to pkg/openstorage
